feat: refactor admin layout to use Suspense and separate client logic
